Move BaseRequestFuture to the right package
